requirements.txt
pycryptodome
pwntools